Fixed #5. Spider is correctly initialized with settings for testing.
Test setup now initializes a crawler with the settings and binds it to the
spider. This allows code in the spider that set default values for settings
and checked whether the spider was bound to a crawler before accessing any
settings to be removed. Now the settings are accessed directly since they
will always be defined.

Updated the tests for the eBird spider to compare the values in the initialzed
spider with the values from the settings, compared to the magic values used
previously.