Fixed #4.  Settings are now overridden from local_settings.py in the same directory as settings.py
Previously local_settings.py was located in the root directory of the project but that
made it awkward to install in the global environment,a virtualenv or to deploy as an
egg to a scrapyd daemon. Moving local_settings.py into the same directory as the
settings.py (and updating the install docs) solves this problem.