Fixes #15. Protocols are now defined using a name and not a code.
Using codes made assumptions about the final destination where the
checklists would be loaded - specifically the database model in the
django-checklists project. Using the name extracted from the site
(WorldBirds does not define a protocol so a default 'Timed visit'
is used) makes it easier to map protocols to different codes, if
any, when the data is loaded into its destination.