Fixed merging of checklist entries.
When updating the checklist entries extracted from the ebird API with
the entries extracted from the checklist web page problems can occur
when dealing with subspecies. The records from the API only contain
the species while the checklist web page contains the subspecies.
Matching the two is easy if there is only one entry for the species
but when species and subspecies are included and both have the same
counts then it is impossible to differentiate them.

The methods used to merge the checklists now take this into account
so the entry is not merged - only the entry from the API is used.
A warning system was added so any such mismatches can be recorded
and included in the status report sent when the crawler completes.
A warning is also generated when an entry from the checklist web
page does not have a matching entry from the API. This is possibly a
bug in the eBird API and so it's useful to keep track of this.