Passed country code to parsers to use as a prefix for identifiers.
The country code used to identify which WorldBirds database to access
is passed to the various parsers in the request meta-data so it can
be used as a prefix when generating the identifiers for checklists
and locations. This allows data from different WorldBirds databases
to be aggregated without the identifiers (which are almost certainly
primary keys) colliding.