Improved logging when tracking down records missing from the API.
The eBird API does not always return all the records for a given
location. Often the web page used to display the checklist on the
eBird web site has between 1 and 20 (approx) additional records.
Improved the logging to record the URLs for the API call and the
web page to make it easier to compare and identify which records
were available only through the web page.