Renamed project and package to checklists_scrapers.
Since the project is intended to be used to create feeds of checklists
for a django-checklists database the project we renamed to make the
association clearer.

Renamed environment variables used to define the values for settings
to use the "CHECKLISTS" prefix.

The mail server settings can be set from environment variables that
match the name of the scrapy settings, e.g. CHECKLISTS_MAIL_HOST to
set the MAIL_HOST and also from the equivalent environment variables
used in the django-checklists project, e.g. CHECKLISTS_EMAIL_HOST.
This reduces the number of environment variables that need to be set
when using the checklists_scrapers and django-checklists projects
together.

Changed the licence for the project to BSD, from the GPLv3, so it is
compatible with the other projects in the checklists "group".