Counts broken down by age and sex now contain an identifier.
The table on the web page for a checklist on eBird contains a table
showing a breakdown of the count by age and sex. These counts are
extracted and added as detail records to the corresponding entry.
In order to uniquely identfy the count and so track any changes made
after the checklist was first downloaded each cell in the table is
given an identifer calculated from the row and column number (1 based).