Rounded coordinates to 4 decimal places.
The coordinates found in WorldBirds appear to have some form of s
systematic error such that they are expressed to 13 decimal places
e.g.  38.684, -9.3451666666667. When the coordinates are extracted
they are rounded to the nearest 4 decimal places so that they are
presented with a reasonable amount of precision and to ensure that
the values don't overflow the fields in the checklists database.