Added language selector and fixed parsing checklist attributes.
Added an initial request to select the language (English) to display
the web pages in. Although the language is listed in the preferences
for the account being used it is not applied when logging in.

Fixed the parsing of the checklist attributes. Instead of relying
on fixed row indices (which depended on the data entered) the values
are extracted from the table using the label.